third_party/blink/web_tests/svg/dynamic-updates/SVGRadialGradientElement-svgdom-cy-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cy' property of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.cy.baseVal.valueAsString is "0%"
PASS radialGradientElement.cy.baseVal.valueAsString is "150%"
PASS successfullyParsed is true

TEST COMPLETE
